fix: Use internal test configuration with minimal web environment
